User commit

+++ src/main/resources/entity/Student.java
@@ -1,12 +1,12 @@
 package com.example.demo.entity;
 public class Student {
-    private int id ;
+    private long id ;
     private String name;
     private String email;
     public void setId(int id){
         this.id=id;
     }
-    public int getId() {
+    public long getId() {
         return id;
     }
     public String getName() {
@@ -21,7 +21,7 @@ public class Student {
     public void setEmail(String email) {
         this.email = email;
     }
-    public Student(int id, String name, String email) {
+    public Student(long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
